Put FEM 'packing' functions into separate header (#3621)
* Simplify coefficient packing for Expressions

* Work on packing

* Merge pack functions

* Simplify

* Simplify

* Break up fem/utils.h

* Add pack.h

* Add concept

* Fix docstring

* Simplifications

* Simplifications

* Small edit
